Move metrics_of_subject() to external.database.reports.
